refactor: почистил, поперемещал, описал иерархию контекстов
