+++ src/bsl/scope/localModuleScope.ts
@@ -1,15 +1,14 @@
-import { BaseScope, Scope } from './scope'
-import { Symbol, SymbolType } from './symbols'
+import { BaseScope, Scope, SymbolType } from '@/scope'
 import { editor } from "monaco-editor-core"
-import { parse } from "../bsl/parser"
-import { Method, Module } from "../bsl/Symbols"
+import { parse } from "../parser"
+import { Method, Module } from "../Symbols"
+import { createMethodScope } from './methodScope'
 
 export class LocalModuleScope extends BaseScope {
     private readonly model: editor.ITextModel
     private module: Module = {
         vars: [], methods: []
     }
-
 
     constructor(model: editor.ITextModel) {
         super([])
@@ -31,22 +30,7 @@ export class LocalModuleScope extends BaseScope {
         if (method === undefined) {
             return undefined
         }
-
-        const members: Symbol[] = []
-        method.autoVars.forEach(v => members.push({
-            name: v.name,
-            kind: SymbolType.property,
-        }))
-        method.vars.forEach(v => members.push({
-            name: v.name,
-            kind: SymbolType.property,
-        }))
-        method.params.forEach(v => members.push({
-            name: v.name,
-            kind: SymbolType.property,
-        }))
-
-        return new BaseScope(members)
+        return createMethodScope(method)
     }
 
     private needUpdateMembers(): boolean {
