+++ src/scope/index.ts
@@ -2,5 +2,5 @@ export * from './symbols'
 export * from './scope'
 export * from './types'
 export * from './globalScopeManager'
-export * from './editorScope'
-export * from './localModuleScope'
+export * from '../bsl/scope/editorScope'
+export * from '../bsl/scope/localModuleScope'
