+++ src/yaxunit/interfaces.ts
@@ -1,0 +1,5 @@
+import { TestsModel } from "./test-model";
+
+export interface TestModelRender{
+    update(model: TestsModel): void
+}
